WIP: refactoring of derived model module
