+++ resqpy/derived_model.py
@@ -35,1119 +35,15 @@ import resqpy.rq_import as rqi
 import resqpy.well as rqw
 
 
-def _pl(n, use_es = False):
-    if n == 1:
-        return ''
-    elif use_es:
-        return 'es'
-    else:
-        return 's'
 
 
-def _prepare_simple_inheritance(grid, source_grid, inherit_properties, inherit_realization, inherit_all_realizations):
-    collection = None
-    if inherit_properties:
-        source_collection = source_grid.extract_property_collection()
-        if source_collection is not None:
-            #  do not inherit the inactive property array by this mechanism
-            collection = rqp.GridPropertyCollection()
-            collection.set_grid(grid)
-            collection.extend_imported_list_copying_properties_from_other_grid_collection(
-                source_collection, realization = inherit_realization, copy_all_realizations = inherit_all_realizations)
-    return collection
 
 
-def zone_layer_ranges_from_array(zone_array, min_k0 = 0, max_k0 = None, use_dominant_zone = False):
-    """Returns a list of (zone_min_k0, zone_max_k0, zone_index) derived from zone_array.
 
-    arguments:
-       zone_array (3D numpy int array or masked array): array holding zone index value per cell
-       min_k0 (int, default 0): the minimum layer number (0 based) to be included in the ranges
-       max_k0 (int, default None): the maximum layer number (0 based) to be included in the ranges;
-          note that this layer is included (unlike in python ranges); if None, the maximum layer
-          number in zone_array is used
-       use_dominant_zone (boolean, default False): if True, the most common zone value in each layer is used for the whole
-          layer; if False, then variation of zone values in active cells in a layer will raise an assertion error
 
-    returns:
-       a list of (int, int, int) being (zone_min_k0, zone_max_k0, zone_index) for each zone index value present
 
-    notes:
-       the function requires zone indices (for active cells, if zone_array is masked) within a layer
-       to be consistent: an assertion error is raised otherwise; the returned list is sorted by layer
-       ranges rather than zone index; if use_dominant_zone is True then a side effect of the function is
-       to modify the values in zone_array to be consistent across each layer, effectively reassigning some
-       cells to a different zone!
-    """
 
-    def dominant_zone(zone_array):
-        # modifies data in zone_array such that each layer has a single (most common) value
-        for k in range(zone_array.shape[0]):
-            unique_zones = np.unique(zone_array[k])
-            if len(unique_zones) <= 1:
-                continue
-            dominant = unique_zones[0]
-            dominant_count = np.count_nonzero(zone_array[k] == dominant)
-            for contender in unique_zones[1:]:
-                contender_count = np.count_nonzero(zone_array[k] == contender)
-                if contender_count > dominant_count:
-                    dominant = contender
-                    dominant_count = contender_count
-            zone_array[k] = dominant
-            log.info('layer ' + str(k + 1) + ' (1 based) zone set to most common value ' + str(dominant))
 
-    if max_k0 is None:
-        max_k0 = zone_array.shape[0] - 1
-    if use_dominant_zone:
-        dominant_zone(zone_array)
-    zone_list = np.unique(zone_array)
-    log.debug('list of zones: ' + str(zone_list))
-    assert len(zone_list) > 0, 'no zone values present (all cells inactive?)'
-    zone_layer_range_list = []  # list of (zone_min_k0, zone_max_k0, zone)
-    for zone in zone_list:
-        single_zone_mask = (zone_array == zone)
-        zone_min_k0 = None
-        zone_max_k0 = None
-        for k0 in range(min_k0, max_k0 + 1):
-            if np.any(single_zone_mask[k0]):
-                zone_min_k0 = k0
-                break
-        if zone_min_k0 is None:
-            log.warning('no active cells for zone ' + str(zone) + ' in layer range ' + str(min_k0 + 1) + ' to ' +
-                        str(max_k0 + 1))
-            continue
-        for k0 in range(max_k0, min_k0 - 1, -1):
-            if np.any(single_zone_mask[k0]):
-                zone_max_k0 = k0
-                break
-        # require all active cells in zone layer range to be for this zone
-        for k0 in range(zone_min_k0, zone_max_k0 + 1):
-            assert np.all(single_zone_mask[k0]), 'unacceptable zone variation with layer ' + str(k0 + 1)
-        zone_layer_range_list.append((zone_min_k0, zone_max_k0, zone))
-    assert len(zone_layer_range_list) > 0, 'no zone layer ranges derived from zone array'
-    zone_layer_range_list.sort()
-    for zone_i in range(1, len(zone_layer_range_list)):
-        assert zone_layer_range_list[zone_i][0] > zone_layer_range_list[
-            zone_i - 1][1], 'overlapping zone layer ranges'  # todo: add more info
-        if zone_layer_range_list[zone_i][0] > zone_layer_range_list[zone_i - 1][1] + 1:
-            log.warning('gap in zonal layer ranges, missing layer(s) being arbitrarily assigned to zone below'
-                       )  # todo: add more info to log
-            zone_layer_range_list[zone_i][0] = zone_layer_range_list[zone_i - 1][1] + 1
-    return zone_layer_range_list
-
-
-def add_zone_by_layer_property(epc_file,
-                               grid_uuid = None,
-                               zone_by_layer_vector = None,
-                               zone_by_cell_property_uuid = None,
-                               use_dominant_zone = False,
-                               use_local_property_kind = True,
-                               null_value = -1,
-                               title = 'ZONE',
-                               realization = None,
-                               extra_metadata = {}):
-    """Adds a discrete zone property (and local property kind) with indexable element of layers.
-
-    arguments:
-       epc_file (string): file name to load resqml model from and to update with the zonal property
-       grid_uuid (uuid.UUID or str, optional): required unless the model has only one grid, or one named ROOT
-       zone_by_layer_vector (nk integers, optional): either this or zone_by_cell_property_uuid must be given;
-          a 1D numpy array, tuple or list of ints, being the zone number to which each layer belongs
-       zone_by_cell_property_uuid (uuid.UUID or str, optional): either this or zone_by_layer_vector must be given;
-          the uuid of a discrete property with grid as supporting representation and cells as indexable elements,
-          holidng the zone to which the cell belongs
-       use_dominant_zone (boolean, default False): if True and more than one zone is represented within the cells
-          of a layer, then the whole layer is assigned to the zone with the biggest count of cells in the layer;
-          if False, an exception is raised if more than one zone is represented by the cells of a layer; ignored
-          if zone_by_cell_property_uuid is None
-       use_local_property_kind (boolean, default True): if True, the new zone by layer property is given a
-          local property kind titled 'zone'; if False, the property kind will be set to 'discrete'
-       null_value (int, default -1): the value to use if a layer does not belong to any zone (rarely used)
-       title (str, default 'ZONE'): the citation title of the new zone by layer property
-       realization (int, optional): if present the new zone by layer property is marked as belonging to this
-          realization
-       extra_metadata (dict, optional): any items in this dictionary are added as extra metadata to the new
-          property
-
-    returns:
-       numpy vector of zone numbers (by layer), uuid of newly created property
-    """
-
-    assert zone_by_layer_vector is not None or zone_by_cell_property_uuid is not None
-    assert zone_by_layer_vector is None or zone_by_cell_property_uuid is None
-
-    model = rq.Model(epc_file)
-
-    if grid_uuid is None:
-        grid = model.grid()
-        grid_uuid = grid.uuid
-    else:
-        grid = model.grid_for_uuid_from_grid_list(grid_uuid)
-        if grid is None:
-            grid = grr.any_grid(model, uuid = grid_uuid, find_properties = True)
-    assert grid is not None, 'failed to establish grid object'
-
-    if zone_by_layer_vector is not None:
-        assert len(
-            zone_by_layer_vector) == grid.nk, 'length of zone by layer vector does not match number of layers in grid'
-        zone_by_layer = np.array(zone_by_layer_vector, dtype = int)
-    elif zone_by_cell_property_uuid is not None:
-        pc = grid.property_collection
-        zone_by_cell_array = pc.single_array_ref(uuid = zone_by_cell_property_uuid)
-        assert zone_by_cell_array is not None, 'zone by cell property array not found for uuid: ' + str(
-            zone_by_cell_property_uuid)
-        zone_range_list = zone_layer_ranges_from_array(zone_by_cell_array, use_dominant_zone = use_dominant_zone)
-        assert zone_range_list is not None and len(
-            zone_range_list) > 0, 'failed to convert zone by cell to zone by layer'
-        zone_by_layer = np.full((grid.nk,), null_value, dtype = int)
-        for min_k0, max_k0, zone_index in zone_range_list:
-            assert 0 <= min_k0 <= max_k0 < grid.nk, 'zonal layer limits out of range for grid (probable bug)'
-            zone_by_layer[min_k0:max_k0 + 1] = zone_index
-    else:
-        raise Exception('code failure')
-
-    assert zone_by_layer.ndim == 1 and zone_by_layer.size == grid.nk
-
-    if use_local_property_kind:
-        property_kind = 'zone'
-        zone_pk = grr.establish_zone_property_kind(model)
-        local_property_kind_uuid = zone_pk.uuid
-        model.store_epc(only_if_modified = True)
-    else:
-        property_kind = 'discrete'
-        local_property_kind_uuid = None
-
-    model.h5_release()
-
-    property_uuid = add_one_grid_property_array(epc_file,
-                                                zone_by_layer,
-                                                property_kind,
-                                                grid_uuid = grid.uuid,
-                                                source_info = 'zone defined by layer',
-                                                title = title,
-                                                discrete = True,
-                                                null_value = null_value,
-                                                indexable_element = 'layers',
-                                                realization = realization,
-                                                local_property_kind_uuid = local_property_kind_uuid,
-                                                extra_metadata = extra_metadata)
-
-    return zone_by_layer, property_uuid
-
-
-def add_one_grid_property_array(epc_file,
-                                a,
-                                property_kind,
-                                grid_uuid = None,
-                                source_info = 'imported',
-                                title = None,
-                                discrete = False,
-                                uom = None,
-                                time_index = None,
-                                time_series_uuid = None,
-                                string_lookup_uuid = None,
-                                null_value = None,
-                                indexable_element = 'cells',
-                                facet_type = None,
-                                facet = None,
-                                realization = None,
-                                local_property_kind_uuid = None,
-                                count_per_element = 1,
-                                const_value = None,
-                                points = False,
-                                extra_metadata = {},
-                                new_epc_file = None):
-    """Adds a grid property from a numpy array to an existing resqml dataset.
-
-    arguments:
-       epc_file (string): file name to load model resqml model from (and rewrite to if new_epc_file is None)
-       a (3D numpy array): the property array to be added to the model; for a constant array set this None
-          and use the const_value argument, otherwise this array is required
-       property_kind (string): the resqml property kind
-       grid_uuid (uuid object or string, optional): the uuid of the grid to which the property relates;
-          if None, the property is attached to the 'main' grid
-       source_info (string): typically the name of a file from which the array has been read but can be any
-          information regarding the source of the data
-       title (string): this will be used as the citation title when a part is generated for the array; for simulation
-          models it is desirable to use the simulation keyword when appropriate
-       discrete (boolean, default False): if True, the array should contain integer (or boolean) data; if False, float
-       uom (string, default None): the resqml units of measure for the data; not relevant to discrete data
-       time_index (integer, default None): if not None, the time index to be used when creating a part for the array
-       time_series_uuid (uuid object or string, default None): required if time_index is not None
-       string_lookup_uuid (uuid object or string, optional): required if the array is to be stored as a categorical
-          property; set to None for non-categorical discrete data; only relevant if discrete is True
-       null_value (int, default None): if present, this is used in the metadata to indicate that this value
-          is to be interpreted as a null value wherever it appears in the data (use for discrete data only)
-       indexable_element (string, default 'cells'): the indexable element in the supporting representation (the grid)
-       facet_type (string): resqml facet type, or None
-       facet (string): resqml facet, or None
-       realization (int): realization number, or None
-       local_property_kind_uuid (uuid.UUID or string): uuid of local property kind, or None
-       count_per_element (int, default 1): the number of values per indexable element; if greater than one then this
-          must be the fastest cycling axis in the cached array, ie last index
-       const_value (float or int, optional): if present, a constant array is added 'filled' with this value, in which
-          case argument a should be None
-       points (bool, default False): if True, this is a points property with an extra dimension of extent 3
-       extra_metadata (dict, optional): any items in this dictionary are added as extra metadata to the new
-          property
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new property object; if present,
-          a new epc file (& associated h5 file) is created to contain a copy of the grid and the new property
-
-    returns:
-       uuid.UUID of newly created property object
-    """
-
-    if new_epc_file and epc_file and (
-        (new_epc_file == epc_file) or
-        (os.path.exists(new_epc_file) and os.path.exists(epc_file) and os.path.samefile(new_epc_file, epc_file))):
-        new_epc_file = None
-
-    # open up model and establish grid object
-    model = rq.Model(epc_file)
-    if grid_uuid is None:
-        grid = model.grid()
-        grid_uuid = grid.uuid
-    else:
-        grid = model.grid_for_uuid_from_grid_list(grid_uuid)
-        if grid is None:
-            grid = grr.any_grid(model, uuid = grid_uuid, find_properties = False)
-    assert grid is not None, 'failed to establish grid object'
-
-    if not discrete:
-        string_lookup_uuid = None
-
-    # create an empty property collection and add the new array to its 'imported' list
-    gpc = rqp.GridPropertyCollection()
-    gpc.set_grid(grid)
-    gpc.add_cached_array_to_imported_list(a,
-                                          source_info,
-                                          title,
-                                          discrete = discrete,
-                                          uom = uom,
-                                          time_index = time_index,
-                                          null_value = null_value,
-                                          property_kind = property_kind,
-                                          local_property_kind_uuid = local_property_kind_uuid,
-                                          facet_type = facet_type,
-                                          facet = facet,
-                                          realization = realization,
-                                          indexable_element = indexable_element,
-                                          count = count_per_element,
-                                          const_value = const_value,
-                                          points = points)
-
-    # write or re-write model
-    model.h5_release()
-    if new_epc_file:
-        grid_title = rqet.citation_title_for_node(grid.root)
-        uuid_list = write_grid(new_epc_file,
-                               grid,
-                               property_collection = gpc,
-                               grid_title = grid_title,
-                               mode = 'w',
-                               time_series_uuid = time_series_uuid,
-                               string_lookup_uuid = string_lookup_uuid,
-                               extra_metadata = extra_metadata)
-    else:
-        # add arrays to hdf5 file holding source grid geometry
-        uuid_list = write_grid(epc_file,
-                               grid,
-                               property_collection = gpc,
-                               mode = 'a',
-                               geometry = False,
-                               time_series_uuid = time_series_uuid,
-                               string_lookup_uuid = string_lookup_uuid,
-                               extra_metadata = extra_metadata)
-
-    if uuid_list is None or len(uuid_list) == 0:
-        return None
-    return uuid_list[0]
-
-
-def add_one_blocked_well_property(epc_file,
-                                  a,
-                                  property_kind,
-                                  blocked_well_uuid,
-                                  source_info = 'imported',
-                                  title = None,
-                                  discrete = False,
-                                  uom = None,
-                                  time_index = None,
-                                  time_series_uuid = None,
-                                  string_lookup_uuid = None,
-                                  null_value = None,
-                                  indexable_element = 'cells',
-                                  facet_type = None,
-                                  facet = None,
-                                  realization = None,
-                                  local_property_kind_uuid = None,
-                                  count_per_element = 1,
-                                  points = False,
-                                  extra_metadata = {},
-                                  new_epc_file = None):
-    """Adds a blocked well property from a numpy array to an existing resqml dataset.
-
-    arguments:
-       epc_file (string): file name to load model resqml model from (and rewrite to if new_epc_file is None)
-       a (1D numpy array): the blocked well property array to be added to the model
-       property_kind (string): the resqml property kind
-       blocked_well_uuid (uuid object or string): the uuid of the blocked well to which the property relates
-       source_info (string): typically the name of a file from which the array has been read but can be any
-          information regarding the source of the data
-       title (string): this will be used as the citation title when a part is generated for the array
-       discrete (boolean, default False): if True, the array should contain integer (or boolean) data; if False, float
-       uom (string, default None): the resqml units of measure for the data; not relevant to discrete data
-       time_index (integer, default None): if not None, the time index to be used when creating a part for the array
-       time_series_uuid (uuid object or string, default None): required if time_index is not None
-       string_lookup_uuid (uuid object or string, optional): required if the array is to be stored as a categorical
-          property; set to None for non-categorical discrete data; only relevant if discrete is True
-       null_value (int, default None): if present, this is used in the metadata to indicate that this value
-          is to be interpreted as a null value wherever it appears in the data (use for discrete data only)
-       indexable_element (string, default 'cells'): the indexable element in the supporting representation (the blocked well);
-          valid values are 'cells', 'intervals' (which includes unblocked intervals), or 'nodes'
-       facet_type (string): resqml facet type, or None
-       facet (string): resqml facet, or None
-       realization (int): realization number, or None
-       local_property_kind_uuid (uuid.UUID or string): uuid of local property kind, or None
-       count_per_element (int, default 1): the number of values per indexable element; if greater than one then this
-          must be the fastest cycling axis in the cached array, ie last index; if greater than 1 then a must be a 2D array
-       points (bool, default False): if True, this is a points property with an extra dimension of extent 3
-       extra_metadata (dict, optional): any items in this dictionary are added as extra metadata to the new
-          property
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new property object; if present,
-          a new epc file (& associated h5 file) is created to contain a copy of the blocked well (and dependencies) and
-          the new property
-
-    returns:
-       uuid.UUID of newly created property object
-    """
-
-    if new_epc_file and epc_file and (
-        (new_epc_file == epc_file) or
-        (os.path.exists(new_epc_file) and os.path.exists(epc_file) and os.path.samefile(new_epc_file, epc_file))):
-        new_epc_file = None
-
-    # open up model and establish grid object
-    model = rq.Model(epc_file)
-
-    blocked_well = rqw.BlockedWell(model, uuid = blocked_well_uuid)
-    assert blocked_well is not None, f'no blocked well object found with uuid {blocked_well_uuid}'
-
-    if not discrete:
-        string_lookup_uuid = None
-
-    # create an empty property collection and add the new array to its 'imported' list
-    bwpc = rqp.PropertyCollection()
-    bwpc.set_support(support = blocked_well, model = model)
-    bwpc.add_cached_array_to_imported_list(a,
-                                           source_info,
-                                           title,
-                                           discrete = discrete,
-                                           uom = uom,
-                                           time_index = time_index,
-                                           null_value = null_value,
-                                           property_kind = property_kind,
-                                           local_property_kind_uuid = local_property_kind_uuid,
-                                           facet_type = facet_type,
-                                           facet = facet,
-                                           realization = realization,
-                                           indexable_element = indexable_element,
-                                           count = count_per_element,
-                                           points = points)
-    bwpc.write_hdf5_for_imported_list()
-    uuid_list = bwpc.create_xml_for_imported_list_and_add_parts_to_model(time_series_uuid = time_series_uuid,
-                                                                         string_lookup_uuid = string_lookup_uuid,
-                                                                         property_kind_uuid = local_property_kind_uuid,
-                                                                         extra_metadata = extra_metadata)
-    assert len(uuid_list) == 1
-    model.store_epc()
-    return uuid_list[0]
-
-
-def add_wells_from_ascii_file(epc_file,
-                              crs_uuid,
-                              trajectory_file,
-                              comment_character = '#',
-                              space_separated_instead_of_csv = False,
-                              well_col = 'WELL',
-                              md_col = 'MD',
-                              x_col = 'X',
-                              y_col = 'Y',
-                              z_col = 'Z',
-                              length_uom = 'm',
-                              md_domain = None,
-                              drilled = False,
-                              z_inc_down = True,
-                              new_epc_file = None):
-    """Adds new md datum, trajectory, interpretation and feature objects for each well in a tabular ascii file..
-
-    arguments:
-       epc_file (string): file name to load model resqml model from (and rewrite to if new_epc_file is None)
-       crs_uuid (uuid.UUID): the unique identifier of the coordinate reference system applicable to the x,y,z data;
-          if None, a default crs will be created, making use of the length_uom and z_inc_down arguments
-       trajectory_file (string): the path of the ascii file holding the well trajectory data to be loaded
-       comment_character (string, default '#'): character deemed to introduce a comment in the trajectory file
-       space_separated_instead_of_csv (boolean, default False): if True, the columns in the trajectory file are space
-          separated; if False, comma separated
-       well_col (string, default 'WELL'): the heading for the column containing well names
-       md_col (string, default 'MD'): the heading for the column containing measured depths
-       x_col (string, default 'X'): the heading for the column containing X (usually easting) data
-       y_col (string, default 'Y'): the heading for the column containing Y (usually northing) data
-       z_col (string, default 'Z'): the heading for the column containing Z (depth or elevation) data
-       length_uom (string, default 'm'): the units of measure for the measured depths; should be 'm' or 'ft'
-       md_domain (string, optional): the source of the original deviation data; may be 'logger' or 'driller'
-       drilled (boolean, default False): True should be used for wells that have been drilled; False otherwise (planned,
-          proposed, or a location being studied)
-       z_inc_down (boolean, default True): indicates whether z values increase with depth; only used in the creation
-          of a default coordinate reference system; ignored if crs_uuid is not None
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new property object; if present,
-          a new epc file (& associated h5 file) is created to contain a copy of the grid and the new property
-
-    returns:
-       int: the number of wells added
-
-    notes:
-       ascii file must be table with first line being column headers, with columns for WELL, MD, X, Y & Z;
-       actual column names can be set with optional arguments;
-       all the objects are added to the model, with array data being written to the hdf5 file for the trajectories;
-       the md_domain and drilled values are stored in the RESQML metadata but are only for human information and do not
-       generally affect computations
-    """
-
-    assert trajectory_file and os.path.exists(trajectory_file)
-    if md_domain:
-        assert md_domain in ['driller', 'logger']
-
-    if new_epc_file and epc_file and (
-        (new_epc_file == epc_file) or
-        (os.path.exists(new_epc_file) and os.path.exists(epc_file) and os.path.samefile(new_epc_file, epc_file))):
-        new_epc_file = None
-
-    # open up model
-    if new_epc_file:
-        model = rq.Model(new_epc_file, copy_from = epc_file)
-    else:
-        model = rq.Model(epc_file)
-
-    # sort out the coordinate reference system
-    if crs_uuid is None:
-        crs_uuid = rqet.uuid_for_part_root(model.crs_root)
-    if crs_uuid is None:
-        if z_inc_down is None:
-            z_inc_down = True
-        crs_root = rqcrs.Crs(model, xy_units = length_uom, z_units = length_uom, z_inc_down = z_inc_down)
-        crs_uuid = rqet.uuid_for_part_root(crs_root)
-
-    # add all the well related objects to the model, based on data in the ascii file
-    (feature_list, interpretation_list, trajectory_list, md_datum_list) =  \
-       rqw.add_wells_from_ascii_file(model, crs_uuid, trajectory_file, comment_character = comment_character,
-                                     space_separated_instead_of_csv = space_separated_instead_of_csv,
-                                     well_col = well_col, md_col = md_col, x_col = x_col, y_col = y_col, z_col = z_col,
-                                     length_uom = length_uom, md_domain = md_domain, drilled = drilled)
-
-    assert len(feature_list) == len(interpretation_list) == len(trajectory_list) == len(md_datum_list)
-    count = len(feature_list)
-
-    log.info('features, interpretations, trajectories and md data added for ' + str(count) + ' well' + _pl(count))
-
-    # write or re-write model
-    model.h5_release()
-    model.store_epc()
-
-    return count
-
-
-def zonal_grid(epc_file,
-               source_grid = None,
-               zone_title = None,
-               zone_uuid = None,
-               zone_layer_range_list = None,
-               k0_min = None,
-               k0_max = None,
-               use_dominant_zone = False,
-               inactive_laissez_faire = True,
-               new_grid_title = None,
-               new_epc_file = None):
-    """Extends an existing model with a new version of the source grid converted to a single, thick, layer per zone.
-
-    arguments:
-       epc_file (string): file name to rewrite the model's xml to; if source grid is None, model is loaded from this file
-       source_grid (grid.Grid object, optional): a multi-layer RESQML grid object; if None, the epc_file is loaded
-          and it should contain one ijk grid object (or one 'ROOT' grid) which is used as the source grid
-       zone_title (string): if not None, a discrete property with this as the citation title is used as the zone property
-       zone_uuid (string or uuid): if not None, a discrete property with this uuid is used as the zone property (see notes)
-       zone_layer_range_list (list of (int, int, int)): each entry being (min_k0, max_k0, zone_index); alternative to
-          working from a zone array
-       k0_min (int, optional): the minimum layer number in the source grid (zero based) to include in the zonal version;
-          default is zero (ie. top layer in source grid)
-       k0_max (int, optional): the maximum layer number in the source grid (zero based) to include in the zonal version;
-          default is nk - 1 (ie. bottom layer in source grid)
-       use_dominant_zone (boolean, default False): if True, the most common zone value in each layer is used for the whole
-          layer; if False, then variation of zone values in active cells in a layer will raise an assertion error
-       inactive_laissez_faire (boolean, optional): if True, a cell in the zonal grid will be set active if any of the
-          corresponding cells in the source grid are active; otherwise all corresponding cells in the source grid
-          must be active for the zonal cell to be active; default is True
-       new_grid_title (string): used as the citation title text for the new grid object
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new grid object; if present,
-          a new epc file (& associated h5 file) is created to contain the zonal grid (& crs)
-
-    returns:
-       new grid object (grid.Grid) with one layer per zone of the source grid
-
-    notes:
-       usually one of zone_title or zone_uuid or zone_layer_range_list should be passed, if none are passed then a
-       single layer grid is generated; zone_layer_range_list will take precendence if present
-    """
-
-    def fetch_zone_array(grid, zone_title = None, zone_uuid = None, masked = True):
-        properties = grid.extract_property_collection()
-        assert properties is not None and properties.number_of_parts() > 0, 'no properties found in relation to grid'
-        properties = rqp.selective_version_of_collection(properties, continuous = False)
-        assert properties is not None and properties.number_of_parts() > 0,  \
-           'no discreet properties found in relation to grid'
-        if zone_title:
-            properties = rqp.selective_version_of_collection(
-                properties, citation_title = zone_title)  # could make case insensitive?
-            assert properties is not None and properties.number_of_parts() > 0,  \
-               'no discreet property found with title ' + zone_title
-        if zone_uuid:
-            if isinstance(zone_uuid, str):
-                zone_uuid = bu.uuid_from_string(zone_uuid)
-            zone_uuid_str = str(zone_uuid)
-            if zone_title:
-                postamble = ' (and title ' + zone_title + ')'
-            else:
-                postamble = ''
-            assert zone_uuid in properties.uuids(), 'no property found with uuid ' + zone_uuid_str + postamble
-            part_name = grid.model.part(uuid = zone_uuid)
-        else:
-            part_name = properties.singleton()
-        return properties.cached_part_array_ref(part_name, masked = masked)  # .copy() needed?
-
-    assert epc_file or new_epc_file, 'epc file name not specified'
-    if new_epc_file and epc_file and (
-        (new_epc_file == epc_file) or
-        (os.path.exists(new_epc_file) and os.path.exists(epc_file) and os.path.samefile(new_epc_file, epc_file))):
-        new_epc_file = None
-    assert epc_file or source_grid is not None, 'neither epc file name nor source grid supplied'
-    if source_grid is None:
-        model = rq.Model(epc_file)
-        source_grid = model.grid()  # requires there to be exactly one grid in model (or one named ROOT)
-    else:
-        model = source_grid.model
-    assert source_grid.grid_representation in ['IjkGrid', 'IjkBlockGrid']
-    if source_grid.grid_representation == 'IjkBlockGrid':
-        source_grid.make_regular_points_cached()
-    assert model is not None
-    single_layer_mode = (not zone_title and not zone_uuid)
-
-    if k0_min is None:
-        k0_min = 0
-    else:
-        assert k0_min >= 0 and k0_min < source_grid.nk
-    if k0_max is None:
-        k0_max = source_grid.nk - 1
-    else:
-        assert k0_max >= 0 and k0_max < source_grid.nk and k0_max >= k0_min
-
-    if not single_layer_mode:  # process zone array
-        if zone_layer_range_list is None:
-            zone_array = fetch_zone_array(source_grid, zone_title, zone_uuid)
-            zone_layer_range_list = zone_layer_ranges_from_array(zone_array,
-                                                                 k0_min,
-                                                                 k0_max,
-                                                                 use_dominant_zone = use_dominant_zone)
-        zone_count = len(zone_layer_range_list)
-        # above is list of (zone_min_k0, zone_max_k0, zone) sorted by zone_min_k0
-        log.info('following layer ranges are based on top layer being numbered 1 (simulator protocol)')
-        for (zone_min_k0, zone_max_k0, zone) in zone_layer_range_list:
-            log.info('zone id {0:1d} covers layers {1:1d} to {2:1d}'.format(zone, zone_min_k0 + 1, zone_max_k0 + 1))
-    else:
-        zone_layer_range_list = [(k0_min, k0_max, 0)]
-        zone_count = 1
-    assert zone_count > 0, 'unexpected lack of zones'
-
-    # create a new, empty grid object
-    is_regular = grr.is_regular_grid(source_grid.root) and single_layer_mode
-    if is_regular:
-        dxyz_dkji = source_grid.block_dxyz_dkji.copy()
-        dxyz_dkji[0] *= k0_max - k0_min + 1
-        grid = grr.RegularGrid(model,
-                               extent_kji = (1, source_grid.nj, source_grid.ni),
-                               dxyz_dkji = dxyz_dkji,
-                               origin = source_grid.block_origin,
-                               crs_uuid = source_grid.crs_uuid,
-                               set_points_cached = False)
-    else:
-        grid = grr.Grid(model)
-        # inherit attributes from source grid
-        grid.grid_representation = 'IjkGrid'
-        grid.extent_kji = np.array((zone_count, source_grid.nj, source_grid.ni), dtype = 'int')
-        grid.nk, grid.nj, grid.ni = zone_count, source_grid.nj, source_grid.ni
-        grid.has_split_coordinate_lines = source_grid.has_split_coordinate_lines
-        grid.crs_root = source_grid.crs_root
-        grid.crs_uuid = source_grid.crs_uuid
-
-    grid.k_direction_is_down = source_grid.k_direction_is_down
-    grid.grid_is_right_handed = source_grid.grid_is_right_handed
-    grid.pillar_shape = source_grid.pillar_shape
-
-    # aggregate inactive cell mask depending on laissez faire argument
-    if source_grid.inactive is None:
-        log.debug('setting inactive mask to None')
-        grid.inactive = None
-    elif single_layer_mode:
-        if inactive_laissez_faire:
-            log.debug('setting inactive mask using all mode (laissez faire)')
-            grid.inactive = np.all(source_grid.inactive[k0_min:k0_max + 1], axis = 0).reshape(grid.extent_kji)
-        else:
-            log.debug('setting inactive mask using any mode (strict)')
-            grid.inactive = np.any(source_grid.inactive[k0_min:k0_max + 1], axis = 0).reshape(grid.extent_kji)
-    else:
-        grid.inactive = np.zeros(grid.extent_kji, dtype = bool)
-        for zone_i in range(zone_count):
-            zk0_min, zk0_max, _ = zone_layer_range_list[zone_i]
-            if inactive_laissez_faire:
-                grid.inactive[zone_i] = np.all(source_grid.inactive[zk0_min:zk0_max + 1], axis = 0)
-            else:
-                grid.inactive[zone_i] = np.any(source_grid.inactive[zk0_min:zk0_max + 1], axis = 0)
-
-    if not is_regular:
-
-        # rework the grid geometry
-        source_grid.cache_all_geometry_arrays()
-        # determine cell geometry is defined
-        if hasattr(source_grid, 'array_cell_geometry_is_defined'):
-            grid.array_cell_geometry_is_defined = np.empty(grid.extent_kji, dtype = bool)
-            if single_layer_mode:
-                grid.array_cell_geometry_is_defined[0] = np.logical_and(
-                    source_grid.array_cell_geometry_is_defined[k0_min],
-                    source_grid.array_cell_geometry_is_defined[k0_max])
-            else:
-                for zone_i in range(zone_count):
-                    zk0_min, zk0_max, _ = zone_layer_range_list[zone_i]
-                    grid.array_cell_geometry_is_defined[zone_i] = np.logical_and(
-                        source_grid.array_cell_geometry_is_defined[zk0_min],
-                        source_grid.array_cell_geometry_is_defined[zk0_max])
-                    # could attempt to pick up some corner points from K-neighbouring cells
-            grid.geometry_defined_for_all_cells_cached = np.all(grid.array_cell_geometry_is_defined)
-        else:
-            grid.geometry_defined_for_all_cells_cached = source_grid.geometry_defined_for_all_cells_cached
-        # copy info for pillar geometry is defined
-        grid.geometry_defined_for_all_pillars_cached = source_grid.geometry_defined_for_all_pillars_cached
-        if hasattr(source_grid, 'array_pillar_geometry_is_defined'):
-            grid.array_pillar_geometry_is_defined = source_grid.array_pillar_geometry_is_defined.copy()
-        # get reference to points for source grid geometry
-        source_points = source_grid.points_ref()
-        # slice top and base points
-        points_shape = list(source_points.shape)
-        points_shape[0] = zone_count + 1  # nk + 1
-        grid.points_cached = np.zeros(points_shape)
-        if grid.geometry_defined_for_all_cells_cached:
-            if single_layer_mode:
-                grid.points_cached[0] = source_points[k0_min]
-                grid.points_cached[1] = source_points[k0_max + 1]  # base face
-            else:
-                for zone_i in range(zone_count):
-                    if zone_i == 0:
-                        grid.points_cached[0] = source_points[zone_layer_range_list[zone_i][0]]
-                    grid.points_cached[zone_i + 1] = source_points[
-                        zone_layer_range_list[zone_i][1]]  # or could use 0th element of tuple for zone_i+1
-        elif not grid.has_split_coordinate_lines:
-            log.debug('scanning columns (unsplit pillars) for reference geometry')
-            # fill in geometry: todo: replace with array operations if possible
-            for zone_i in range(zone_count):
-                zk0_min, zk0_max = zone_layer_range_list[zone_i][0:2]
-                for j in range(grid.nj):
-                    for i in range(grid.ni):
-                        if zone_i == 0:
-                            for k in range(zk0_min, zk0_max + 1):
-                                if source_grid.array_cell_geometry_is_defined[k, j, i]:
-                                    grid.points_cached[0, j:j + 2, i:i + 2] = source_points[k, j:j + 2, i:i + 2]
-                                    break
-                        for k in range(zk0_max, zk0_min - 1, -1):
-                            if source_grid.array_cell_geometry_is_defined[k, j, i]:
-                                grid.points_cached[zone_i + 1, j:j + 2, i:i + 2] = source_points[k + 1, j:j + 2,
-                                                                                                 i:i + 2]
-                                grid.array_cell_geometry_is_defined[zone_i, j, i] = True
-                                break
-        else:
-            log.debug('scanning columns (split pillars) for reference geometry')
-            if not hasattr(source_grid, 'pillars_for_column'):
-                source_grid.create_column_pillar_mapping()
-            grid.pillars_for_column = source_grid.pillars_for_column.copy()
-            for zone_i in range(zone_count):
-                zk0_min, zk0_max = zone_layer_range_list[zone_i][0:2]
-                for j in range(grid.nj):
-                    for i in range(grid.ni):
-                        if grid.inactive[zone_i, j, i]:
-                            continue
-                        if zone_i == 0:
-                            for k in range(zk0_min, zk0_max + 1):
-                                if source_grid.array_cell_geometry_is_defined[k, j, i]:
-                                    for jp in range(2):
-                                        for ip in range(2):
-                                            pillar = grid.pillars_for_column[j, i, jp, ip]
-                                            grid.points_cached[0, pillar] = source_points[k, pillar]
-                                    break
-                        for k in range(zk0_max + 1, zk0_min - 1, -1):
-                            if source_grid.array_cell_geometry_is_defined[k, j, i]:
-                                for jp in range(2):
-                                    for ip in range(2):
-                                        pillar = grid.pillars_for_column[j, i, jp, ip]
-                                        grid.points_cached[zone_i + 1, pillar] = source_points[k + 1, pillar]
-                                grid.array_cell_geometry_is_defined[zone_i, j, i] = True
-                                break
-        if grid.has_split_coordinate_lines:
-            grid.split_pillar_indices_cached = source_grid.split_pillar_indices_cached.copy()
-            grid.cols_for_split_pillars = source_grid.cols_for_split_pillars.copy()
-            grid.cols_for_split_pillars_cl = source_grid.cols_for_split_pillars_cl.copy()
-            grid.split_pillars_count = source_grid.split_pillars_count
-
-    if new_grid_title is None or len(new_grid_title) == 0:
-        if single_layer_mode:
-            preamble = 'single layer'
-        else:
-            preamble = 'zonal'
-        new_grid_title = preamble + ' version of ' + str(rqet.citation_title_for_node(source_grid.root))
-
-    model.h5_release()
-    if new_epc_file:
-        write_grid(new_epc_file, grid, grid_title = new_grid_title, mode = 'w')
-    else:
-        #      ext_uuid, _ = model.h5_uuid_and_path_for_node(rqet.find_nested_tags(source_grid.root, ['Geometry', 'Points']), 'Coordinates')
-        write_grid(epc_file, grid, ext_uuid = None, grid_title = new_grid_title, mode = 'a')
-
-    return grid
-
-
-def single_layer_grid(epc_file,
-                      source_grid = None,
-                      k0_min = None,
-                      k0_max = None,
-                      inactive_laissez_faire = True,
-                      new_grid_title = None,
-                      new_epc_file = None):
-    """Extends an existing model with a new version of the source grid converted to a single, thick, layer.
-
-    arguments:
-       epc_file (string): file name to rewrite the model's xml to; if source grid is None, model is loaded from this file
-       source_grid (grid.Grid object, optional): a multi-layer RESQML grid object; if None, the epc_file is loaded
-          and it should contain one ijk grid object (or one 'ROOT' grid) which is used as the source grid
-       k0_min (int, optional): the minimum layer number in the source grid (zero based) to include in the single layer version;
-          default is zero (ie. top layer in source grid)
-       k0_max (int, optional): the maximum layer number in the source grid (zero based) to include in the single layer version;
-          default is nk - 1 (ie. bottom layer in source grid)
-       inactive_laissez_faire (boolean, optional): if True, a cell in the single layer grid will be set active if any of the
-          corresponding cells in the source grid are active; otherwise all corresponding cells in the source grid
-          must be active for the single layer cell to be active; default is True
-       new_grid_title (string): used as the citation title text for the new grid object
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new grid object; if present,
-          a new epc file (& associated h5 file) is created to contain the single layer grid (& crs)
-
-    returns:
-       new grid object (grid.Grid) with a single layer representation of the source grid
-    """
-
-    return zonal_grid(epc_file,
-                      source_grid = source_grid,
-                      k0_min = k0_min,
-                      k0_max = k0_max,
-                      inactive_laissez_faire = inactive_laissez_faire,
-                      new_grid_title = new_grid_title,
-                      new_epc_file = new_epc_file)
-
-
-def interpolated_grid(epc_file,
-                      grid_a,
-                      grid_b,
-                      a_to_b_0_to_1 = 0.5,
-                      split_tolerance = 0.01,
-                      inherit_properties = False,
-                      inherit_realization = None,
-                      inherit_all_realizations = False,
-                      new_grid_title = None,
-                      new_epc_file = None):
-    """Extends an existing model with a new grid geometry linearly interpolated between the two source_grids.
-
-    arguments:
-       epc_file (string): file name to rewrite the model's xml to
-       grid_a, grid_b (grid.Grid objects): a pair of RESQML grid objects representing the end cases, between
-          which the new grid will be interpolated
-       a_to_b_0_to_1 (float, default 0.5): the interpolation factor in the range zero to one; a value of 0.0 will yield
-          a copy of grid a, a value of 1.0 will yield a copy of grid b, intermediate values will yield a grid with all
-          points interpolated
-       split_tolerance (float, default 0.01): maximum offset of corner points for shared point to be generated; units
-          are same as those in grid crs; only relevant if working from corner points, ignored otherwise
-       inherit_properties (boolean, default False): if True, the new grid will have a copy of any properties associated
-          with grid_a
-       inherit_realization (int, optional): realization number for which properties will be inherited; ignored if
-          inherit_properties is False
-       inherit_all_realizations (boolean, default False): if True (and inherit_realization is None), properties for all
-          realizations will be inherited; if False, only properties with a realization of None are inherited; ignored if
-          inherit_properties is False or inherit_realization is not None
-       new_grid_title (string): used as the citation title text for the new grid object
-       new_epc_file (string, optional): if None, the source epc_file is extended with the new grid object; if present,
-          a new epc file (& associated h5 file) is created to contain the interpolated grid (& crs)
-
-    returns:
-       new grid object (grid.Grid) with geometry interpolated between grid a and grid b
-
-    notes:
-       the hdf5 file used by the grid_a model is appended to, so it is recommended that the grid_a model's epc is specified
-       as the first argument (unless a new epc file is required, sharing the hdf5 file)
-    """
-
-    assert epc_file or new_epc_file, 'epc file name not specified'
-    if new_epc_file and epc_file and (
-        (new_epc_file == epc_file) or
-        (os.path.exists(new_epc_file) and os.path.exists(epc_file) and os.path.samefile(new_epc_file, epc_file))):
-        new_epc_file = None
-    assert grid_a is not None and grid_b is not None, 'at least one source grid is missing'
-    assert grid_a.grid_representation == 'IjkGrid' and grid_b.grid_representation == 'IjkGrid'
-    assert 0.0 <= a_to_b_0_to_1 <= 1.0, 'interpolation factor outside range 0.0 to 1.0'
-    assert tuple(grid_a.extent_kji) == tuple(grid_b.extent_kji), 'source grids have different extents'
-    assert grid_a.k_direction_is_down == grid_b.k_direction_is_down, 'source grids have different k directions'
-    assert grid_a.grid_is_right_handed == grid_b.grid_is_right_handed, 'source grids have different ijk handedness'
-    assert grid_a.pillar_shape == grid_b.pillar_shape, 'source grids have different resqml pillar shapes'
-
-    b_weight = a_to_b_0_to_1
-    a_weight = 1.0 - b_weight
-
-    model = grid_a.model
-
-    if not bu.matching_uuids(grid_a.crs_uuid, grid_b.crs_uuid):
-        crs_a = rqcrs.Crs(grid_a.model, uuid = grid_a.crs_uuid)
-        crs_b = rqcrs.Crs(grid_b.model, uuid = grid_b.crs_uuid)
-        assert crs_a.is_equivalent(
-            crs_b), 'end point grids for interpolation have different coordinate reference systems'
-
-    log.info('loading geometry for two source grids')
-    grid_a.cache_all_geometry_arrays()
-    grid_b.cache_all_geometry_arrays()
-
-    assert grid_a.geometry_defined_for_all_cells() and grid_b.geometry_defined_for_all_cells(
-    ), 'geometry not defined for all cells'
-    #  assert grid_a.geometry_defined_for_all_pillars() and grid_b.geometry_defined_for_all_pillars(), 'geometry not defined for all pillars'
-
-    if not grid_a.has_split_coordinate_lines and not grid_b.has_split_coordinate_lines:
-        work_from_pillars = True
-    elif (grid_a.has_split_coordinate_lines and grid_b.has_split_coordinate_lines and
-          grid_a.points_cached.shape == grid_b.points_cached.shape and
-          grid_a.split_pillar_indices_cached.shape == grid_b.split_pillar_indices_cached.shape and
-          grid_a.cols_for_split_pillars.shape == grid_b.cols_for_split_pillars.shape and
-          grid_a.cols_for_split_pillars_cl.shape == grid_b.cols_for_split_pillars_cl.shape and
-          np.all(grid_a.split_pillar_indices_cached == grid_b.split_pillar_indices_cached) and
-          np.all(grid_a.cols_for_split_pillars == grid_b.cols_for_split_pillars) and
-          np.all(grid_a.cols_for_split_pillars_cl == grid_b.cols_for_split_pillars_cl)):
-        work_from_pillars = True
-    else:
-        work_from_pillars = False
-
-    if work_from_pillars:
-        log.info('interpolating between compatible pillar grids')
-    else:
-        log.warning('interpolating between corner points due to pillar incompatibilities')
-
-    # create a new, empty grid object
-    grid = grr.Grid(model)
-
-    # inherit attributes from source grid
-    grid.grid_representation = 'IjkGrid'
-    grid.extent_kji = grid_a.extent_kji.copy()
-    grid.nk, grid.nj, grid.ni = grid.extent_kji
-    grid.k_direction_is_down = grid_a.k_direction_is_down
-    grid.grid_is_right_handed = grid_a.grid_is_right_handed
-    grid.pillar_shape = grid_a.pillar_shape
-    grid.has_split_coordinate_lines = (grid_a.has_split_coordinate_lines or grid_b.has_split_coordinate_lines)
-    # inherit the coordinate reference system used by the grid geometry
-    grid.crs_root = grid_a.crs_root
-    grid.crs_uuid = grid_a.crs_uuid
-
-    if grid_a.inactive is None or grid_b.inactive is None:
-        grid.inactive = None
-    else:
-        grid.inactive = np.logical_and(grid_a.inactive, grid_b.inactive)
-
-    grid.geometry_defined_for_all_cells_cached = True
-    grid.array_cell_geometry_is_defined = np.ones(tuple(grid.extent_kji), dtype = bool)
-    grid.geometry_defined_for_all_pillars_cached = True
-
-    if work_from_pillars:
-        grid.points_cached = grid_a.points_cached * a_weight + grid_b.points_cached * b_weight
-        grid.has_split_coordinate_lines = grid_a.has_split_coordinate_lines
-        if grid.has_split_coordinate_lines:
-            grid.split_pillar_indices_cached = grid_a.split_pillar_indices_cached.copy()
-            grid.cols_for_split_pillars = grid_a.cols_for_split_pillars.copy()
-            grid.cols_for_split_pillars_cl = grid_a.cols_for_split_pillars_cl.copy()
-            grid.split_pillars_count = grid_a.split_pillars_count
-    else:
-        grid.pillar_shape = 'curved'  # following fesapi approach of non-parametric pillars even if they are in fact straight
-        cp_a = grid_a.corner_points(cache_cp_array = True)
-        cp_b = grid_b.corner_points(cache_cp_array = True)
-        assert cp_a.shape == cp_b.shape
-        grid_cp = cp_a * a_weight + cp_b * b_weight
-        if grid.nk > 1:
-            z_gap = grid_cp[1:, :, :, 0, :, :, :] - grid_cp[:-1, :, :, 1, :, :, :]
-            max_gap = np.max(np.abs(z_gap))
-            log.info('maximum vertical void distance after corner point interpolation: {0:.3f} {1}'.format(
-                max_gap, grid_a.z_units()))
-            # close vertical voids (includes shifting x & y)
-            if max_gap > 0.0:
-                log.debug('closing vertical voids')
-                z_gap *= 0.5
-                grid_cp[1:, :, :, 0, :, :, :] -= z_gap
-                grid_cp[:-1, :, :, 1, :, :, :] += z_gap
-        # reduce cp array extent in k
-        log.debug('reducing k extent of interpolated corner point array (sharing points vertically)')
-        k_reduced_cp_array = np.zeros((grid.nk + 1, grid.nj, grid.ni, 2, 2, 3))  # (nk+1, nj, ni, jp, ip, xyz)
-        k_reduced_cp_array[0, :, :, :, :, :] = grid_cp[0, :, :, 0, :, :, :]
-        k_reduced_cp_array[-1, :, :, :, :, :] = grid_cp[-1, :, :, 1, :, :, :]
-        if grid.nk > 1:
-            k_reduced_cp_array[1:-1, :, :, :, :, :] = grid_cp[:-1, :, :, 1, :, :, :]
-        # create primary pillar reference indices as one of four column corners around pillar, active column preferred
-        log.debug('creating primary pillar reference neighbourly indices')
-        primary_pillar_jip = np.zeros((grid.nj + 1, grid.ni + 1, 2), dtype = 'int')  # (nj + 1, ni + 1, jp:ip)
-        primary_pillar_jip[-1, :, 0] = 1
-        primary_pillar_jip[:, -1, 1] = 1
-        # build extra pillar references for split pillars
-        extras_count = np.zeros((grid.nj + 1, grid.ni + 1), dtype = 'int')  # count (0 to 3) of extras for pillar
-        extras_list_index = np.zeros((grid.nj + 1, grid.ni + 1), dtype = 'int')  # index in list of 1st extra for pillar
-        extras_list = []  # list of (jp, ip)
-        extras_use = np.negative(np.ones((grid.nj, grid.ni, 2, 2),
-                                         dtype = 'int'))  # (j, i, jp, ip); -1 means use primary
-        log.debug('building extra pillar references for split pillars')
-        # loop over pillars
-        for j in range(grid.nj + 1):
-            for i in range(grid.ni + 1):
-                primary_jp = primary_pillar_jip[j, i, 0]
-                primary_ip = primary_pillar_jip[j, i, 1]
-                p_col_j = j - primary_jp
-                p_col_i = i - primary_ip
-                # loop over 4 columns surrounding this pillar
-                for jp in range(2):
-                    col_j = j - jp
-                    if col_j < 0 or col_j >= grid.nj:
-                        continue  # no column this side of pillar in j
-                    for ip in range(2):
-                        col_i = i - ip
-                        if col_i < 0 or col_i >= grid.ni:
-                            continue  # no column this side of pillar in i
-                        if jp == primary_jp and ip == primary_ip:
-                            continue  # this column is the primary for this pillar
-                        discrepancy = np.max(
-                            np.abs(k_reduced_cp_array[:, col_j, col_i, jp, ip, :] -
-                                   k_reduced_cp_array[:, p_col_j, p_col_i, primary_jp, primary_ip, :]))
-                        if discrepancy <= split_tolerance:
-                            continue  # data for this column's corner aligns with primary
-                        for e in range(extras_count[j, i]):
-                            eli = extras_list_index[j, i] + e
-                            pillar_j_extra = j - extras_list[eli][0]
-                            pillar_i_extra = i - extras_list[eli][1]
-                            discrepancy = np.max(
-                                np.abs(k_reduced_cp_array[:, col_j, col_i, jp, ip, :] -
-                                       k_reduced_cp_array[:, pillar_j_extra, pillar_i_extra, extras_list[eli][0],
-                                                          extras_list[eli][1], :]))
-                            if discrepancy <= split_tolerance:  # data for this corner aligns with existing extra
-                                extras_use[col_j, col_i, jp, ip] = e
-                                break
-                        if extras_use[col_j, col_i, jp, ip] >= 0:  # reusing an existing extra for this pillar
-                            continue
-                        # add this corner as an extra
-                        if extras_count[j, i] == 0:  # create entry point for this pillar in extras
-                            extras_list_index[j, i] = len(extras_list)
-                        extras_list.append((jp, ip))
-                        extras_use[col_j, col_i, jp, ip] = extras_count[j, i]
-                        extras_count[j, i] += 1
-        if len(extras_list) == 0:
-            grid.has_split_coordinate_lines = False
-        log.debug('number of extra pillars: ' + str(len(extras_list)))
-        # create points array as used in resqml
-        log.debug('creating points array as used in resqml format')
-        if grid.has_split_coordinate_lines:
-            points_array = np.zeros((grid.nk + 1, (grid.nj + 1) * (grid.ni + 1) + len(extras_list), 3))
-            index = 0
-            # primary pillars
-            for pillar_j in range(grid.nj + 1):
-                for pillar_i in range(grid.ni + 1):
-                    (jp, ip) = primary_pillar_jip[pillar_j, pillar_i]
-                    points_array[:, index, :] = k_reduced_cp_array[:, pillar_j - jp, pillar_i - ip, jp, ip, :]
-                    index += 1
-            # add extras for split pillars
-            for pillar_j in range(grid.nj + 1):
-                for pillar_i in range(grid.ni + 1):
-                    for e in range(extras_count[pillar_j, pillar_i]):
-                        eli = extras_list_index[pillar_j, pillar_i] + e
-                        (jp, ip) = extras_list[eli]
-                        pillar_j_extra = pillar_j - jp
-                        pillar_i_extra = pillar_i - ip
-                        points_array[:, index, :] = k_reduced_cp_array[:, pillar_j_extra, pillar_i_extra, jp, ip, :]
-                        index += 1
-            assert index == (grid.nj + 1) * (grid.ni + 1) + len(extras_list)
-        else:  # unsplit pillars
-            points_array = np.zeros((grid.nk + 1, grid.nj + 1, grid.ni + 1, 3))
-            for j in range(grid.nj + 1):
-                for i in range(grid.ni + 1):
-                    (jp, ip) = primary_pillar_jip[j, i]
-                    points_array[:, j, i, :] = k_reduced_cp_array[:, j - jp, i - ip, jp, ip, :]
-        grid.points_cached = points_array
-        # add split pillar arrays to grid object
-        if grid.has_split_coordinate_lines:
-            log.debug('adding split pillar arrays to grid object')
-            split_pillar_indices_list = []
-            cumulative_length_list = []
-            cols_for_extra_pillar_list = []
-            cumulative_length = 0
-            for pillar_j in range(grid.nj + 1):
-                for pillar_i in range(grid.ni + 1):
-                    for e in range(extras_count[pillar_j, pillar_i]):
-                        split_pillar_indices_list.append(pillar_j * (grid.ni + 1) + pillar_i)
-                        use_count = 0
-                        for jp in range(2):
-                            j = pillar_j - jp
-                            if j < 0 or j >= grid.nj:
-                                continue
-                            for ip in range(2):
-                                i = pillar_i - ip
-                                if i < 0 or i >= grid.ni:
-                                    continue
-                                if extras_use[j, i, jp, ip] == e:
-                                    use_count += 1
-                                    cols_for_extra_pillar_list.append((j * grid.ni) + i)
-                        assert (use_count > 0)
-                        cumulative_length += use_count
-                        cumulative_length_list.append(cumulative_length)
-            log.debug('number of extra pillars: ' + str(len(split_pillar_indices_list)))
-            assert (len(cumulative_length_list) == len(split_pillar_indices_list))
-            grid.split_pillar_indices_cached = np.array(split_pillar_indices_list, dtype = 'int')
-            log.debug('number of uses of extra pillars: ' + str(len(cols_for_extra_pillar_list)))
-            assert (len(cols_for_extra_pillar_list) == np.count_nonzero(extras_use + 1))
-            assert (len(cols_for_extra_pillar_list) == cumulative_length)
-            grid.cols_for_split_pillars = np.array(cols_for_extra_pillar_list, dtype = 'int')
-            assert (len(cumulative_length_list) == len(extras_list))
-            grid.cols_for_split_pillars_cl = np.array(cumulative_length_list, dtype = 'int')
-            grid.split_pillars_count = len(extras_list)
-
-    collection = _prepare_simple_inheritance(grid, grid_a, inherit_properties, inherit_realization,
-                                             inherit_all_realizations)
-
-    if new_grid_title is None or len(new_grid_title) == 0:
-        new_grid_title = 'interpolated between two grids with factor: ' + str(a_to_b_0_to_1)
-
-    model.h5_release()
-    if new_epc_file:
-        write_grid(new_epc_file, grid, property_collection = collection, grid_title = new_grid_title, mode = 'w')
-    else:
-        ext_uuid, _ = model.h5_uuid_and_path_for_node(rqet.find_nested_tags(grid_a.root, ['Geometry', 'Points']),
-                                                      'Coordinates')
-        write_grid(epc_file,
-                   grid,
-                   ext_uuid = ext_uuid,
-                   property_collection = collection,
-                   grid_title = new_grid_title,
-                   mode = 'a')
-
-    return grid
 
 
 def extract_box(epc_file = None,
@@ -3687,126 +2583,7 @@ def displacement_properties(new_grid, old_grid):
     return displacement_collection
 
 
-def write_grid(epc_file,
-               grid,
-               ext_uuid = None,
-               property_collection = None,
-               grid_title = None,
-               mode = 'a',
-               geometry = True,
-               time_series_uuid = None,
-               string_lookup_uuid = None,
-               extra_metadata = {}):
-    """Append to or create epc and h5 files, with grid and optionally property collection.
 
-    arguments:
-       epc_file (string): name of existing epc file (if appending) or new epc file to be created (if writing)
-       grid (grid.Grid object): the grid object to be written to the epc & h5 files
-       ext_uuid (uuid.UUID object, optional): if present and the mode is 'a', the arrays are appended to
-          the hdf5 file that has this uuid; if None or mode is 'w', the hdf5 file is determined automatically
-       property_collection (property.GridPropertyCollection object, optional): if present, a collection of
-          grid properties to write and relate to the grid
-       grid_title (string): used as the citation title for the grid object
-       mode (string, default 'a'): 'a' or 'w'; if 'a', epc_file should be an existing file which is extended
-          (appended to) with the new grid and properties; if 'w', epc_file is created along with an h5 file
-          and will be populated with the grid, crs and properties
-       geometry (boolean, default True): if True, the grid object is included in the write; if False, only the
-          property collection is written, in which case grid must be fully established with xml in place
-       time_series_uuid (uuid.UUID, optional): the uuid of a time series object; required if property_collection
-          contains any recurrent properties in its import list
-       string_lookup_uuid (optional): if present, the uuid of the string table lookup which any non-continuous
-          properties relate to (ie. they are all taken to be categorical); leave as None if discrete property
-          objects are required rather than categorical
-       extra_metadata (dict, optional): any items in this dictionary are added as extra metadata to any new
-          properties
-
-    returns:
-       list of uuid.UUID, being the uuids of property parts added from the property_collection, if any
-
-    note:
-       this function is not usually called directly by application code
-    """
-
-    log.debug('write_grid(): epc_file: ' + str(epc_file) + '; mode: ' + str(mode) + '; grid extent: ' +
-              str(grid.extent_kji))
-
-    assert mode in ['a', 'w']
-    assert geometry or mode == 'a', 'for now, a copy of the grid must be included with property collection'
-    assert geometry or property_collection is not None, 'write_grid called without anything to write'
-
-    if not epc_file.endswith('.epc'):
-        epc_file += '.epc'
-
-    is_regular = grr.is_regular_grid(grid.root)
-
-    if not is_regular:
-        grid.cache_all_geometry_arrays()
-    working_model = grid.model
-
-    if mode == 'a':  # append to existing model
-        log.debug('write_grid(): re-using existing model object')
-        model = working_model
-        if ext_uuid is None:
-            ext_uuid = model.h5_uuid()
-    else:  # create new model with crs, grid and properties
-        log.debug('write_grid(): creating new model object')
-        model = rq.new_model(epc_file)
-        ext_uuid = model.h5_uuid()
-        crs_root = model.duplicate_node(grid.crs_root)
-        grid.model = model
-        grid.crs_root = crs_root
-    log.debug('write_grid(): number of starting parts: ' + str(model.number_of_parts()))
-
-    if grid.inactive is not None and geometry:
-        inactive_count = np.count_nonzero(grid.inactive)
-        if inactive_count == grid.inactive.size:
-            log.warning('writing grid with all cells inactive')
-        else:
-            log.info(f'grid has {grid.inactive.size - inactive_count} active cells out of {grid.inactive.size}')
-    collection = property_collection
-    if collection is not None:
-        log.debug('write_grid(): number of properties in collection: ' + str(collection.number_of_parts()))
-
-    # append to hdf5 file using arrays cached in grid above
-    hdf5_file = model.h5_file_name(uuid = ext_uuid, file_must_exist = (mode == 'a'))
-    if geometry or collection is not None:
-        log.debug('writing grid arrays to hdf5 file')
-        grid.write_hdf5_from_caches(hdf5_file, mode = mode, geometry = geometry, imported_properties = collection)
-        model.h5_release()
-    if ext_uuid is None:
-        ext_uuid = model.h5_uuid()
-
-    # build xml for grid geometry
-    if geometry:
-        log.debug('building xml for grid object')
-        ijk_node = grid.create_xml(ext_uuid, add_as_part = True, add_relationships = True, title = grid_title)
-        assert ijk_node is not None, 'failed to create IjkGrid node in xml tree'
-        if collection is not None:
-            collection.set_grid(grid, grid_root = grid.root)
-        grid.geometry_root = rqet.find_tag(ijk_node, 'Geometry')
-
-    # add derived inactive array as part
-    if collection is not None:
-        prop_uuid_list = collection.create_xml_for_imported_list_and_add_parts_to_model(
-            ext_uuid,
-            support_uuid = grid.uuid,
-            time_series_uuid = time_series_uuid,
-            string_lookup_uuid = string_lookup_uuid,
-            extra_metadata = extra_metadata)
-    else:
-        prop_uuid_list = []
-
-    # set grid related short cuts in Model object
-    if geometry:
-        model.grid_list.append(grid)
-
-    log.debug('write_grid(): number of finishing parts: ' + str(model.number_of_parts()))
-
-    # store new version of model
-    log.info('writing model xml data in epc file: ' + epc_file)
-    model.store_epc(epc_file)
-
-    return prop_uuid_list
 
 
 def add_edges_per_column_property_array(epc_file,
